feat(auth): register logic
