+++ src/lib/auth.ts
@@ -1,9 +1,10 @@
 'use client'
 
-import { useSearchParams } from 'next/navigation'
+import { usePathname, useSearchParams } from 'next/navigation'
 import { z } from 'zod'
 
-import { formSchema } from '@/components/form/login/schema'
+import { formSchema as loginFormSchema } from '@/components/form/login/schema'
+import { formSchema as registerFormSchema } from '@/components/form/register/schema'
 import AuthService from '@/services/clien-side/auth'
 import { useAuthStore } from '@/store/use-auth-store'
 
@@ -12,7 +13,7 @@ export const useLogin = () => {
   const searchParams = useSearchParams()
   const nextHref = searchParams.get('href') || '/'
 
-  const onSubmit = async (values: z.infer<typeof formSchema>) => {
+  const onSubmit = async (values: z.infer<typeof loginFormSchema>) => {
     const res = await AuthService.login(values)
 
     setUser({
@@ -36,4 +37,29 @@ export const useLogout = () => {
   }
 
   return { onLogout }
+}
+
+export const useRegister = () => {
+  const { nextUrl } = useSwitchAuthPages()
+
+  const onSubmit = async (values: z.infer<typeof registerFormSchema>) => {
+    await AuthService.register({
+      email: values.email,
+      password: values.password,
+    })
+
+    window.location.href = nextUrl
+  }
+
+  return { onSubmit }
+}
+
+export const useSwitchAuthPages = () => {
+  const pathname = usePathname()
+  const isLoginPage = pathname.includes('/login')
+  const nextUrl = isLoginPage
+    ? pathname.replace('login', 'register')
+    : pathname.replace('register', 'login')
+
+  return { nextUrl, isLoginPage }
 }
